% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beaware-package.R
\docType{package}
\name{beaware-package}
\alias{beaware}
\alias{beaware-package}
\title{beaware: R Session Tracking Tool For RSworkbench}
\description{
Tracking sessions around users to know what's going on.
}
\author{
\strong{Maintainer}: Cervan Girard \email{cervan@thinkr.fr} (\href{https://orcid.org/0000-0002-4816-4624}{ORCID})

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
  \item Santé publique France [sponsor]
}

}
\keyword{internal}
