% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{clean_vec}
\alias{clean_vec}
\title{clean_vec}
\usage{
clean_vec(
  vec,
  verbose = FALSE,
  unique = TRUE,
  keep_number = FALSE,
  translit = TRUE,
  punct = TRUE
)
}
\arguments{
\item{vec}{vector}

\item{verbose}{TRUE or FALSE}

\item{unique}{make it unique}

\item{keep_number}{keep number}

\item{translit}{TRUE or FALSE}

\item{punct}{check punct}
}
\description{
clean_vec
}
