% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{color_r_version}
\alias{color_r_version}
\title{Get colors for r_version}
\usage{
color_r_version(data)
}
\arguments{
\item{data}{data from get_info_all}
}
\value{
vector of color
}
\description{
Get colors for r_version
}
