% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_info.R
\name{get_all_info}
\alias{get_all_info}
\title{Get all info about meme and cpu rsession}
\usage{
get_all_info()
}
\value{
data.frame
}
\description{
Get all info about meme and cpu rsession
}
\examples{
get_all_info()
get_all_info()
}
