% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{get_all_r_version}
\alias{get_all_r_version}
\title{Get installed R versions on system}
\usage{
get_all_r_version(path = Sys.getenv("PATH_TO_R_INSTALLED", "/opt/R"))
}
\arguments{
\item{path}{Path to R installed, by default '/opt/R'. You can cahnge this path by set up PATH_TO_R_INSTALLED in your Renviron.}
}
\value{
list of R versions
}
\description{
Get installed R versions on system
}
