% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{get_all_users}
\alias{get_all_users}
\title{Get users form system}
\usage{
get_all_users(data)
}
\arguments{
\item{data}{result of get_all_info}
}
\value{
character vector
}
\description{
Get users form system
}
