% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mem_cpu.R
\name{get_mem_cpu}
\alias{get_mem_cpu}
\title{Get cpu and mem info}
\usage{
get_mem_cpu(rsession = TRUE)
}
\arguments{
\item{rsession}{info only for rsession}
}
\value{
data frame of process rsession
}
\description{
Get cpu and mem info
}
\examples{
get_mem_cpu()
}
