% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mem_info.R
\name{get_mem_info_by_user}
\alias{get_mem_info_by_user}
\title{get mem info for users}
\usage{
get_mem_info_by_user(mem_total)
}
\arguments{
\item{mem_total}{total of memory}
}
\value{
data.frame
}
\description{
get mem info for users
}
\examples{
get_mem_info_by_user(get_mem_info()$memtotal)
}
