% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_process_rsession.R
\name{get_process_rsession}
\alias{get_process_rsession}
\title{Get processes rsession}
\usage{
get_process_rsession()
}
\value{
data frame of process rsession
}
\description{
Get processes rsession
}
\examples{
get_process_rsession()
}
