% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_r_version.R
\name{get_r_version}
\alias{get_r_version}
\title{Get r version for rsession}
\usage{
get_r_version(pids)
}
\arguments{
\item{pids}{vector of pids of rsessions}
}
\value{
data frame of process rsession
}
\description{
Get r version for rsession
}
\examples{
pids <- get_process_rsession()
get_r_version(pids$pid)
}
