% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph}
\alias{graph}
\alias{graph_mem}
\alias{graph_cpu}
\title{Graph memory}
\usage{
graph_mem(data)

graph_cpu(data)
}
\arguments{
\item{data}{data of result get_all_info}
}
\value{
ggplot
}
\description{
Graph memory
}
