% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_mem_all}
\alias{graph_mem_all}
\title{Graph usage of memory}
\usage{
graph_mem_all(data, max_mem)
}
\arguments{
\item{data}{data from get_mem_info_by_user}

\item{max_mem}{max of memory}
}
\value{
ggplot
}
\description{
Graph usage of memory
}
