% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{pull_unique}
\alias{pull_unique}
\title{Cet unique entry from var}
\usage{
pull_unique(.data, var)
}
\arguments{
\item{.data}{data}

\item{var}{var to pull}
}
\value{
data.frame
}
\description{
Cet unique entry from var
}
